removing deprecated things, part 1 (does not compile)
