+++ loader/include/Geode/Loader.hpp
@@ -5,7 +5,7 @@
 #include "loader/Log.hpp"
 #include "loader/Mod.hpp"
 #include "loader/ModEvent.hpp"
-#include "loader/Setting.hpp"
+#include "loader/SettingV3.hpp"
 #include "loader/Dirs.hpp"
 
 #include <Geode/DefaultInclude.hpp>
