+++ loader/include/Geode/UI.hpp
@@ -6,7 +6,8 @@
 #include "ui/EnterLayerEvent.hpp"
 #include "ui/BasedButtonSprite.hpp"
 #include "ui/IconButtonSprite.hpp"
-#include "ui/InputNode.hpp"
+#include "ui/Layout.hpp"
+#include "ui/SpacerNode.hpp"
 #include "ui/General.hpp"
 #include "ui/ListView.hpp"
 #include "ui/MDPopup.hpp"
