+++ loader/include/Geode/loader/ModMetadata.hpp
@@ -2,7 +2,6 @@
 
 #include "../utils/Result.hpp"
 #include "../utils/VersionInfo.hpp"
-#include "Setting.hpp"
 #include "Types.hpp"
 
 #include <matjson.hpp>
@@ -124,13 +123,6 @@ namespace geode {
          */
         [[nodiscard]] std::string getName() const;
         /**
-         * The name of the head developer.
-         * If the mod has multiple * developers, this will return the first 
-         * developer in the list.
-         */
-        [[nodiscard, deprecated("Use ModMetadata::getDevelopers() instead")]]
-        std::string getDeveloper() const;
-        /**
          * The developers of this mod
          */
         [[nodiscard]] std::vector<std::string> getDevelopers() const;
@@ -156,11 +148,6 @@ namespace geode {
          */
         [[nodiscard]] std::optional<std::string> getSupportInfo() const;
         /**
-         * Git Repository of the mod
-         */
-        [[nodiscard, deprecated("Use ModMetadata::getLinks instead")]]
-        std::optional<std::string> getRepository() const;
-        /**
          * Get the links (related websites / servers / etc.) for this mod
          */
         ModMetadataLinks getLinks() const;
@@ -180,11 +167,6 @@ namespace geode {
          * Mod spritesheet names
          */
         [[nodiscard]] std::vector<std::string> getSpritesheets() const;
-        /**
-         * Mod settings
-         * @note Not a map because insertion order must be preserved
-         */
-        [[nodiscard, deprecated("Use getSettingsV3")]] std::vector<std::pair<std::string, Setting>> getSettings() const;
         /**
          * Mod settings
          * @note Not a map because insertion order must be preserved
@@ -237,8 +219,6 @@ namespace geode {
         void setDependencies(std::vector<Dependency> const& value);
         void setIncompatibilities(std::vector<Incompatibility> const& value);
         void setSpritesheets(std::vector<std::string> const& value);
-        [[deprecated("This function does NOTHING")]]
-        void setSettings(std::vector<std::pair<std::string, Setting>> const& value);
         void setSettings(std::vector<std::pair<std::string, matjson::Value>> const& value);
         void setTags(std::unordered_set<std::string> const& value);
         void setNeedsEarlyLoad(bool const& value);
