+++ loader/src/loader/SettingNodeV3.cpp
@@ -1,5 +1,4 @@
 #include "SettingNodeV3.hpp"
-#include <Geode/loader/SettingNode.hpp>
 #include <Geode/utils/ColorProvider.hpp>
 #include <Geode/utils/ranges.hpp>
 #include <Geode/loader/Dirs.hpp>
