+++ loader/src/server/Server.hpp
@@ -3,7 +3,6 @@
 #include "Geode/utils/VersionInfo.hpp"
 #include <Geode/DefaultInclude.hpp>
 #include <Geode/utils/web.hpp>
-#include <Geode/loader/SettingEvent.hpp>
 #include <chrono>
 #include <matjson.hpp>
 #include <vector>
